feat: runtime settings storage + proper CI/CD
this change introduces nix as a CI/CD solution that
reproducibly builds the firmware, which is pushed during
CD to the target device via OTA over MQTT. GitHub Actions
are no longer used, buildbot is used for nix pipelines from
now on.

due to the nature of the nix ecosystem and also due to
the artifacts being public, it is not feasable anymore to
hardcode sensitive information inside the firmware. thus
this change does away with compile-time configuration
and introduces runtime settings storage. the settings
are key-value data pairs that are stored on a new small,
"settings" labeled partition on the device's flash memory.
data can be read and written from/to settings during runtime
using the top-level library crate.

it is also possible to generate a settings partition
beforehand using the `settings-generator` crate, which is
useful for initially configuring the device for its first
run. binary compatibility is guaranteed, because:

- the library algorithms account for endianness, and the
underlying data structure is stable
- both the `firmware` and `settings-generator` crate use the
same library implementation (top-level crate)

because of this, it is safe to simply flash the generated
binary blob to the target device's settings partition.