fix(ci): generate a valid OTA image
previously, the OTA binary was generated, then it was
patched by crane to remove references to the cargo vendor
dir and the toolchain. by changing the contents of the file,
the checksums generated by `espflash` were no longer valid,
thus the OTA image being rejected by the target.

this change changes the order of operations such that the
OTA image is generated last, after patching the binaries.