feat: alarm state persistance + alarm parameter configuration
closes #1

this change introduces alarm state persistance by
self-managing an `persisted-alarm-state` setting which
holds the latest alarm state. this ensures the last state
is restored upon reboot. this also inadvertadly fixes a
security vulnerabily outlined in #1.

alarm timeouts are no longer hardcoded, and can be
configured from settings storage.

individual settings can be overwritten from MQTT on the
topic defined by the `settings-topic-prefix` setting. the
message is expected in the following format:

    <key>\0<value>

that is, the key and value (of the setting to be
overwritten) separated by a null-byte.

also, added a "REBOOT" alarm command which may be an attack
surface for DoS, but is the only way to effectively reload
newly overwritten settings for now. it may be desired to
have a reloading mechanism rather than rebooting the whole
system. no promises yet.