chore(fw,ha_types): rename command `ARM_CUSTOM_BYPASS` to `ARM_INSTANT`
this is a breaking change that requires reconfiguring any automations to issue
an `ARM_INSTANT` command from now on.

previously the command word `ARM_CUSTOM_BYPASS` was used to arm the alarm
instantly. this leads to confusion, because Home Assistant uses this command
word to allow arming only specific zones of the alarm. this is a completely
different functionality to what we used it for.

this change was made to accomodate for zone support later on, and this
hopefully enables a smooth transition, instead of changing the behaviour of
this command in a single version. because of this, the capability of supporting
`ARM_CUSTOM_BYPASS` is not advertised to Home Assistant anymore. if per-zone
arming is implemented, this capability can be re-added with its intended
purpose.